move zendesk client types
