+++ src/clients/index.ts
@@ -1,3 +1,4 @@
 export {newMMClient} from './mattermost';
 export {newAppsClient, AppsClient} from './mattermost_proxy_plugin';
-export {newZDClient, ZDClient} from './zendesk';
+export {newZDClient} from './zendesk';
+export {ZDClient} from './zendesk/types';
