+++ src/clients/zendesk/types.ts
@@ -1,8 +1,3 @@
-import zendesk, {ClientOptions} from 'node-zendesk';
-
-import {Routes} from '../utils';
-import {newConfigStore} from '../store';
-
 interface Tickets {
     create(ticket: any): any;
     exportAudit(id: number): any;
@@ -11,6 +6,17 @@ interface Tickets {
 interface Targets {
     create(target: any): any;
     update(id: string, target: any): any;
+}
+
+export type Webhook = {
+    id: string;
+    name: string;
+    endpoint: string;
+    description: string;
+    http_method: string;
+    request_format: string;
+    status: string;
+    subscriptions: string[];
 }
 
 export interface TicketForms {
@@ -61,20 +67,3 @@ export type ZDClientOptions = {
     botAccessToken: string,
     mattermostSiteUrl: string
 }
-
-export const newZDClient = async (zdOptions: ZDClientOptions): Promise<ZDClient> => {
-    const token = zdOptions.oauth2UserAccessToken;
-    if (!token) {
-        throw new Error('Failed to get oauth2 user access_token');
-    }
-    const config = await newConfigStore(zdOptions.botAccessToken, zdOptions.mattermostSiteUrl).getValues();
-    const remoteUri = config.zd_url + Routes.ZD.APIVersion;
-    const options: ClientOptions = {
-        username: '',
-        token,
-        remoteUri,
-        oauth: true,
-    };
-
-    return zendesk.createClient(options) as ZDClient;
-};
